requirement.txt
torch
pillow
numpy 
matplotlib
transformers
scikit-learn
torchvision